+++ core/utils/date.ts
@@ -44,44 +44,3 @@ export function getZonedDateTime(
     return Temporal.PlainDateTime.from(date).toZonedDateTime(timezone);
   }
 }
-
-/**
- * Parse a date/datetime from a basename.
- *
- * Basenames can be prepended with a date (yyyy-mm-dd) or datetime
- * (yyyy-mm-dd-hh-ii-ss) followed by an underscore (_) or hyphen (-).
- */
-export function parseDateFromBasename(
-  basename: string,
-) {
-  const basenameRegex =
-    /^(?<year>\d{4})-(?<month>\d\d)-(?<day>\d\d)(?:-(?<hour>\d\d)-(?<minute>\d\d)(?:-(?<second>\d\d))?)?(?:_|-)(?<basename>.*)/;
-  const basenameParts = basenameRegex.exec(basename)?.groups;
-
-  if (basenameParts) {
-    const {
-      year,
-      month,
-      day,
-      hour = "00",
-      minute = "00",
-      second = "00",
-      basename,
-    } = basenameParts;
-
-    try {
-      const date = parseDate(
-        `${year}-${month}-${day} ${hour}:${minute}:${second}`,
-      );
-
-      return {
-        date,
-        basename,
-      };
-    } catch {
-      throw new Error(
-        `Invalid date: ${basename} (${year}-${month}-${day} ${hour}:${minute}:${second})`,
-      );
-    }
-  }
-}
